requirements.txt
openai
airbyte
requests
markdown
python-dotenv
